safari-600.7-branch/LayoutTests/fast/regions/cssom/get-regions-by-content2-expected.txt

Test for 90759: [CSSRegions] Rename NamedFlow::getRegionsByContentNode to NamedFlow::getRegionsByContent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regionList is non-null.
PASS regionList.length is 0
PASS regionList2 is non-null.
PASS regionList2.length is 0
PASS regionList != regionList2 is true
PASS regionList3 is non-null.
PASS regionList3.length is 0
PASS regionList4 is non-null.
PASS regionList4.length is 1
PASS regionList4.item(0).id is "region"
PASS regionList5 is non-null.
PASS regionList5.length is 0
PASS regionList6 is non-null.
PASS regionList6.length is 1
PASS regionList6.item(0).id is "region"
PASS regionList7 is non-null.
PASS regionList7.length is 0
PASS regionList8 is non-null.
PASS regionList8.length is 0
PASS regionList9 is non-null.
PASS regionList9.length is 1
PASS successfullyParsed is true

TEST COMPLETE
